feat: [refactor] seperating proto from go
