+++ client/rest/handlers/agent.go
