+++ client/rest/handlers/auth.go
