+++ client/rest/handlers/report.go
