+++ client/rest/handlers/user.go
