+++ client/rest/routes/management.go
