+++ client/rest/routes/report.go
