+++ client/rest/service.go
@@ -1,6 +1,6 @@
-package api
+package rest
 
-import "github.com/chopper-c2-framework/c2-chopper/proto"
+import "github.com/chopper-c2-framework/c2-chopper/grpc/proto"
 
 // IMPLEMENTATION OF THE GRPC CLIENT
 type Client struct {
