+++ server/grpc.go
@@ -5,16 +5,16 @@ import (
 	"log"
 
 	Cfg "github.com/chopper-c2-framework/c2-chopper/core/config"
+	"github.com/chopper-c2-framework/c2-chopper/grpc/proto"
+	_ "github.com/chopper-c2-framework/c2-chopper/grpc/proto"
+	handler "github.com/chopper-c2-framework/c2-chopper/server/grpc"
 
 	"crypto/tls"
-
-	"github.com/chopper-c2-framework/c2-chopper/proto"
-
-	"github.com/chopper-c2-framework/c2-chopper/server/internal/handler"
 
 	"net"
 
 	"google.golang.org/grpc"
+
 	"google.golang.org/grpc/credentials"
 )
