+++ server/grpc/team.go
@@ -1,10 +1,10 @@
-package handler
+package grpc
 
 import (
 	context "context"
 	"fmt"
 
-	"github.com/chopper-c2-framework/c2-chopper/proto"
+	"github.com/chopper-c2-framework/c2-chopper/grpc/proto"
 )
 
 type TeamService struct {
